+++ packages/terra-alert/src/terra-dev-site/doc/alert/Alert.1.doc.jsx
@@ -4,7 +4,7 @@ import Markdown from 'terra-markdown';
 import ReadMe from '../../../../docs/README.md';
 import CustomExampleDocs from '../../../../docs/custom-example.md';
 import ActionAndDismissableExampleDocs from '../../../../docs/action-and-dismissable-example.md';
-import { name } from '../../../../package.json';
+import { name, version } from '../../../../package.json';
 
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 // Component Source
@@ -38,6 +38,7 @@ import ActionAndDismissibleExampleSrc from '!raw-loader!../../../../src/terra-de
 const DocPage = () => (
   <DocTemplate
     packageName={name}
+    packageVersion={version}
     readme={ReadMe}
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
     examples={[