+++ packages/terra-form-select/src/terra-dev-site/doc/form-select/SingleSelectField.04.doc.jsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/SingleSelectField.md';
-import { name } from '../../../../package.json';
+import { name, version } from '../../../../package.json';
 
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 // Component Source
@@ -18,6 +18,7 @@ import RequiredSingleSelectFieldExampleSrc from '!raw-loader!../../../../src/ter
 const DocPage = () => (
   <DocTemplate
     packageName={name}
+    packageVersion={version}
     readme={ReadMe}
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
     propsTables={[{