+++ packages/terra-visually-hidden-text/src/terra-dev-site/doc/VisuallyHiddenText/VisuallyHiddenText.1.doc.jsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
-import { name } from '../../../../package.json';
+import { name, version } from '../../../../package.json';
 
 // Component Source
 import VisuallyHiddenTextSrc from '!raw-loader!../../../../src/VisuallyHiddenText';
@@ -17,6 +17,7 @@ import RefCallbackVisuallyHiddenTextSrc from '!raw-loader!../../../../src/terra-
 const DocPage = () => (
   <DocTemplate
     packageName={name}
+    packageVersion={version}
     readme={ReadMe}
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
     examples={[